+++ src/js_list.hpp
@@ -74,6 +74,12 @@ struct ListClass : ClassDefinition<T, realm::js::List<T>, CollectionClass<T>> {
     static void is_valid(ContextType, FunctionType, ObjectType, size_t, const ValueType[], ReturnValue &);
     static void index_of(ContextType, FunctionType, ObjectType, size_t, const ValueType[], ReturnValue &);
 
+    // aggregate functions
+    static void min(ContextType, FunctionType, ObjectType, size_t, const ValueType[], ReturnValue &);
+    static void max(ContextType, FunctionType, ObjectType, size_t, const ValueType[], ReturnValue &);
+    static void sum(ContextType, FunctionType, ObjectType, size_t, const ValueType[], ReturnValue &);
+    static void avg(ContextType, FunctionType, ObjectType, size_t, const ValueType[], ReturnValue &);
+
     // observable
     static void add_listener(ContextType, FunctionType, ObjectType, size_t, const ValueType[], ReturnValue &);
     static void remove_listener(ContextType, FunctionType, ObjectType, size_t, const ValueType[], ReturnValue &);
@@ -92,6 +98,10 @@ struct ListClass : ClassDefinition<T, realm::js::List<T>, CollectionClass<T>> {
         {"sorted", wrap<sorted>},
         {"isValid", wrap<is_valid>},
         {"indexOf", wrap<index_of>},
+        {"min", wrap<min>},
+        {"max", wrap<max>},
+        {"sum", wrap<sum>},
+        {"avg", wrap<avg>},
         {"addListener", wrap<add_listener>},
         {"removeListener", wrap<remove_listener>},
         {"removeAllListeners", wrap<remove_all_listeners>},
@@ -114,6 +124,75 @@ void ListClass<T>::get_length(ContextType, ObjectType object, ReturnValue &retur
     auto list = get_internal<T, ListClass<T>>(object);
     return_value.set((uint32_t)list->size());
 }
+
+template<typename T>
+void ListClass<T>::min(ContextType ctx, FunctionType, ObjectType this_object, size_t argc, const ValueType arguments[], ReturnValue &return_value) {
+    validate_argument_count_at_least(argc, 1);
+    std::string property_name = Value::validated_to_string(ctx, arguments[0]);
+    
+    auto list = get_internal<T, ListClass<T>>(this_object);
+
+    const ObjectSchema& object_schema = list->get_object_schema();
+    const Property* property = object_schema.property_for_name(property_name);
+    if (!property) {
+        throw std::invalid_argument(util::format("No such property: %1", property_name));
+    }
+
+    util::Optional<Mixed> mixed = list->min(property->table_column);
+    return_value.set(Value::from_mixed(ctx, mixed));
+}
+
+template<typename T>
+void ListClass<T>::max(ContextType ctx, FunctionType, ObjectType this_object, size_t argc, const ValueType arguments[], ReturnValue &return_value) {
+    validate_argument_count_at_least(argc, 1);
+    std::string property_name = Value::validated_to_string(ctx, arguments[0]);
+    
+    auto list = get_internal<T, ListClass<T>>(this_object);
+
+    const ObjectSchema& object_schema = list->get_object_schema();
+    const Property* property = object_schema.property_for_name(property_name);
+    if (!property) {
+        throw std::invalid_argument("No such property");
+    }
+
+    util::Optional<Mixed> mixed = list->max(property->table_column);
+    return_value.set(Value::from_mixed(ctx, mixed));
+}
+
+template<typename T>
+void ListClass<T>::sum(ContextType ctx, FunctionType, ObjectType this_object, size_t argc, const ValueType arguments[], ReturnValue &return_value) {
+    validate_argument_count_at_least(argc, 1);
+    std::string property_name = Value::validated_to_string(ctx, arguments[0]);
+    
+    auto list = get_internal<T, ListClass<T>>(this_object);
+
+    const ObjectSchema& object_schema = list->get_object_schema();
+    const Property* property = object_schema.property_for_name(property_name);
+    if (!property) {
+        throw std::invalid_argument("No such property");
+    }
+
+    util::Optional<Mixed> mixed = list->sum(property->table_column);
+    return_value.set(Value::from_mixed(ctx, mixed));
+}
+
+template<typename T>
+void ListClass<T>::avg(ContextType ctx, FunctionType, ObjectType this_object, size_t argc, const ValueType arguments[], ReturnValue &return_value) {
+    validate_argument_count_at_least(argc, 1);
+    std::string property_name = Value::validated_to_string(ctx, arguments[0]);
+    
+    auto list = get_internal<T, ListClass<T>>(this_object);
+
+    const ObjectSchema& object_schema = list->get_object_schema();
+    const Property* property = object_schema.property_for_name(property_name);    
+    if (!property) {
+        throw std::invalid_argument("No such property");
+    }
+
+    util::Optional<Mixed> mixed = list->average(property->table_column);
+    return_value.set(Value::from_mixed(ctx, mixed));
+}
+
 
 template<typename T>
 void ListClass<T>::get_index(ContextType ctx, ObjectType object, uint32_t index, ReturnValue &return_value) {